chore(backend): Moved shared module to util/shared and drop unused
