+++ .changeset/metal-lions-switch.md
@@ -1,0 +1,5 @@
+---
+'@clerk/shared': minor
+---
+
+Introduce `getClerkJsMajorVersionOrTag()`, `getScriptUrl()`, `callWithRetry()`
