+++ .changeset/silent-fishes-change.md
@@ -1,0 +1,5 @@
+---
+'@clerk/backend': minor
+---
+
+Replace utilities with @clerk/shared exports
