+++ packages/backend/src/tokens/keys.ts
@@ -2,9 +2,9 @@ import { API_URL, API_VERSION, MAX_CACHE_LAST_UPDATED_AT_SECONDS } from '../cons
 // DO NOT CHANGE: Runtime needs to be imported as a default export so that we can stub its dependencies with Sinon.js
 // For more information refer to https://sinonjs.org/how-to/stub-dependency/
 import runtime from '../runtime';
-import { callWithRetry } from '../shared';
 import { joinPaths } from '../util/path';
 import { getErrorObjectByCode } from '../util/request';
+import { callWithRetry } from '../util/shared';
 import {
   TokenVerificationError,
   TokenVerificationErrorAction,
