+++ packages/backend/src/tokens/request.ts
@@ -1,7 +1,7 @@
 import { API_URL, API_VERSION, constants } from '../constants';
-import { isDevelopmentFromApiKey, parsePublishableKey } from '../shared';
 import { assertValidSecretKey } from '../util/assertValidSecretKey';
 import { buildRequest, stripAuthorizationHeader } from '../util/IsomorphicRequest';
+import { isDevelopmentFromApiKey, parsePublishableKey } from '../util/shared';
 import type { RequestState } from './authStatus';
 import { AuthErrorReason, interstitial, signedOut, unknownState } from './authStatus';
 import type { TokenCarrier } from './errors';
